add db client interface
